requirements.txt
Django==1.5
South==0.8.1
coverage==3.6
defusedxml==0.4.1
django-tastypie==0.9.15
lxml==3.2.3
mimeparse==0.1.3
model-mommy==1.0
python-dateutil==2.1
selenium==2.33.0
six==1.3.0